# Use `EqualityComparer<T>` instead of non-generic `object.Equals`