--- a/src/QuikGraph/Algorithms/RootedAlgorithmBase.cs
+++ b/src/QuikGraph/Algorithms/RootedAlgorithmBase.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.Diagnostics;
 #if SUPPORTS_AGGRESSIVE_INLINING
 using System.Runtime.CompilerServices;
@@ -64,7 +65,7 @@ namespace QuikGraph.Algorithms
             if (root == null)
                 throw new ArgumentNullException(nameof(root));
 
-            bool changed = !_hasRootVertex || !Equals(_root, root);
+            bool changed = !_hasRootVertex || !EqualityComparer<TVertex>.Default.Equals(_root, root);
             _root = root;
             _hasRootVertex = true;