… src/QuikGraph/Structures/Edges/SEquatableTaggedEdge.cs
… src/QuikGraph/Structures/Edges/SReversedEdge.cs
… src/QuikGraph/Structures/Edges/STaggedEdge.cs
… src/QuikGraph/Structures/Edges/STaggedUndirectedEdge.cs
… src/QuikGraph/Structures/Edges/TaggedEdge.cs
… src/QuikGraph/Structures/Edges/TaggedUndirectedEdge.cs
… src/QuikGraph/Structures/Graphs/AdjacencyGraph.cs
… src/QuikGraph/Structures/Graphs/ArrayAdjacencyGraph.cs
… src/QuikGraph/Structures/Graphs/ArrayBidirectionalGraph.cs
… src/QuikGraph/Structures/Graphs/ArrayUndirectedGraph.cs
… src/QuikGraph/Structures/Graphs/BidirectionalGraph.cs
… src/QuikGraph/Structures/Graphs/CompressedSparseRowGraph.cs
… src/QuikGraph/Structures/Graphs/DelegateIncidenceGraph.cs
… src/QuikGraph/Structures/Graphs/DelegateUndirectedGraph.cs
… src/QuikGraph/Structures/Graphs/DelegateVertexAndEdgeListGraph.cs
… src/QuikGraph/Structures/Graphs/EdgeListGraph.cs
… src/QuikGraph/Structures/Graphs/UndirectedGraph.cs