## Description
Replace calls to `object.Equals(object)`/`Equals(object, object)` with `EqualityComparer<T>.Default.Equals(T, T)` to reduce boxing when `T` is a value type.

This fixes #5 

## Benchmark
I chose `EdgePredecessorRecorderObserver` and `SEquatableEdge` as a representative benchmark, because it replaces
```c#
OnEdgeDiscovered() 
  SEquatableEdge.Equals(object)
    Equals(SEquatableEdge<TVertex> other)
      TVertex.Equals(object)
        TVertex.Equals(TVertex)
      TVertex.Equals(object)
        TVertex.Equals(TVertex)
```

with approximately

```c#
OnEdgeDiscovered()
  Equals(SEquatableEdge<TVertex> other)
    TVertex.Equals(TVertex)
    TVertex.Equals(TVertex)
```

```c#
[SimpleJob(RuntimeMoniker.NetCoreApp31)]
[MemoryDiagnoser]
public class Tests
{
    private AdjacencyGraph<MyStruct, SEquatableEdge<MyStruct>> graph;

    [Params(10, 100, 1_000, 2_000)]
    public int N { get; set; }

    private readonly struct MyStruct : IEquatable<MyStruct>
    {
        public long Prop1 { get; }
        public long Prop2 { get; }
        public long Prop3 { get; }
        public long Prop4 { get; }

        public MyStruct(int i)
        {
            Prop1 = i;
            Prop2 = i;
            Prop3 = i;
            Prop4 = i;
        }

        public bool Equals(MyStruct other)
        {
            return Prop1 == other.Prop1
                && Prop2 == other.Prop2
                && Prop3 == other.Prop3
                && Prop4 == other.Prop4;
        }

        public override bool Equals(object obj) =>
            obj is MyStruct other && Equals(other);

        public override int GetHashCode() => (int)Prop1;
    }

    [GlobalSetup]
    public void GlobalSetup()
    {
        graph = new AdjacencyGraph<MyStruct, SEquatableEdge<MyStruct>>();
        graph.AddVerticesAndEdgeRange(
            Enumerable.Range(0, N)
                .Select(i => new SEquatableEdge<MyStruct>(new MyStruct(i), new MyStruct(i + 1))));
    }

    [Benchmark]
    public object SEquatableEdge()
    {
        var recorder = new EdgePredecessorRecorderObserver<MyStruct, SEquatableEdge<MyStruct>>();

        var dfs = new EdgeDepthFirstSearchAlgorithm<MyStruct, SEquatableEdge<MyStruct>>(graph);
        using (recorder.Attach(dfs))
        {
            dfs.Compute(new MyStruct(0));
        }

        return dfs;
    }
}
```

## Results for `MyStruct` (large struct)
### master
```
|         Method |    N |          Mean |        Error |        StdDev |       Gen 0 |    Gen 1 |    Gen 2 |    Allocated |
|--------------- |----- |--------------:|-------------:|--------------:|------------:|---------:|---------:|-------------:|
| SEquatableEdge |   10 |      12.16 us |     0.091 us |      0.085 us |      9.1400 |        - |        - |     28.03 KB |
| SEquatableEdge |  100 |     394.99 us |     3.344 us |      3.128 us |    344.7266 |        - |        - |   1057.35 KB |
| SEquatableEdge | 1000 |  50,605.98 us |   695.996 us |    616.982 us |  28500.0000 | 100.0000 | 100.0000 |  87930.17 KB |
| SEquatableEdge | 2000 | 284,436.92 us | 5,655.086 us | 10,759.389 us | 112500.0000 | 500.0000 |        - | 347822.91 KB |
```

### PR
```
|         Method |    N |          Mean |       Error |      StdDev |    Gen 0 |    Gen 1 |    Gen 2 |  Allocated |
|--------------- |----- |--------------:|------------:|------------:|---------:|---------:|---------:|-----------:|
| SEquatableEdge |   10 |      8.283 us |   0.0923 us |   0.0863 us |   3.4943 |        - |        - |   10.73 KB |
| SEquatableEdge |  100 |    157.940 us |   2.5098 us |   2.2249 us |  34.6680 |   0.2441 |        - |  107.01 KB |
| SEquatableEdge | 1000 |  9,640.616 us | 191.9573 us | 413.2084 us | 234.3750 | 156.2500 | 156.2500 | 1078.16 KB |
| SEquatableEdge | 2000 | 38,349.891 us | 675.9138 us | 632.2502 us | 461.5385 | 384.6154 | 384.6154 | 2245.41 KB |
```

### Relative gains, calculated as `PR/master`
```
|         Method |    N | Time | Allocated |
|--------------- |----- |-----:|----------:|
| SEquatableEdge |   10 | 0.61 |    0.3828 |
| SEquatableEdge |  100 | 0.38 |    0.1012 |
| SEquatableEdge | 1000 | 0.19 |    0.0123 |
| SEquatableEdge | 2000 | 0.14 |    0.0006 |
```

## Results for `int` (small struct)
### master
```
|         Method |    N |           Mean |         Error |        StdDev |      Gen 0 |    Gen 1 | Gen 2 |    Allocated |
|--------------- |----- |---------------:|--------------:|--------------:|-----------:|---------:|------:|-------------:|
| BeEquivalentTo |   10 |       8.064 us |     0.0833 us |     0.0779 us |     3.8910 |        - |     - |     11.92 KB |
| BeEquivalentTo |  100 |     219.393 us |     1.8957 us |     1.5830 us |   139.4043 |        - |     - |    427.51 KB |
| BeEquivalentTo | 1000 |  23,921.815 us |   461.3458 us |   566.5741 us | 11593.7500 | 125.0000 |     - |  35907.63 KB |
| BeEquivalentTo | 2000 | 123,744.200 us | 1,229.9758 us | 1,150.5201 us | 46200.0000 |        - |     - | 142148.96 KB |
```

### PR
```
|         Method |    N |          Mean |       Error |      StdDev |    Gen 0 |   Gen 1 |   Gen 2 | Allocated |
|--------------- |----- |--------------:|------------:|------------:|---------:|--------:|--------:|----------:|
| BeEquivalentTo |   10 |      5.326 us |   0.0586 us |   0.0520 us |   1.3885 |       - |       - |   4.26 KB |
| BeEquivalentTo |  100 |     90.722 us |   0.9823 us |   0.9189 us |  10.6201 |       - |       - |  32.77 KB |
| BeEquivalentTo | 1000 |  5,170.236 us |  40.5568 us |  37.9369 us |  93.7500 | 23.4375 |       - | 317.99 KB |
| BeEquivalentTo | 2000 | 20,346.336 us | 164.5421 us | 137.4001 us | 156.2500 | 31.2500 | 31.2500 |  656.7 KB |
```

### Relative gains, calculated as `PR/master`
```
|         Method |    N | Time | Allocated |
|--------------- |----- |-----:|----------:|
| SEquatableEdge |   10 | 0.66 |    0.3574 |
| SEquatableEdge |  100 | 0.41 |    0.0767 |
| SEquatableEdge | 1000 | 0.22 |    0.0088 |
| SEquatableEdge | 2000 | 0.16 |    0.0046 |
```